Fix dma buffer size
The dma buffer consits of two parts with each 500 scans. Number_of_cameras is irrelevant here, because e.g. when there are two cameras the interrupt for copying the data from dma to user buffer is fired twice as often per a given count of scans. 1 scan with 2 cameras is like 2 scans from the perspective of the dma buffer and the interrupt.