Add Labview configurations
Add configuration Release-Labview and Debug-Labview. COMPILE_FOR_LABVIEW
is moved from ESLSCDLL.h to the precompiler definitions in the project
settings. This is done to create valid build setups, now Release and
Debug, when no Labview is installed on the computer. From the
configurations Release and Debug the linking to labview.lib is removed.